Added TOH tests and reverted Pancake puzzle to use ints instead of uint8_t

+++ environments/PancakePuzzle.h
@@ -22,7 +22,7 @@ public:
 		for (unsigned int x = 0; x < N; x++)
 			puzzle[x] = x;
 	}
-	uint8_t puzzle[N];
+	int puzzle[N];
 };
 
 template <int N>
